requirements.txt
python-telegram-bot
sentence-transformers
ollama
numpy
python-dotenv
